+++ src/app/onboarding/components/glam/glam-goals-screen/glam-goals-screen.component.scss
